requirements.txt
asgiref==3.7.2
boto3==1.28.6
botocore==1.31.6
Django>=4.2,<5.0
django-environ==0.4.5
django-storages==1.13.2
gunicorn==20.1.0
jmespath==1.0.1
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2023.3
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
urllib3==2.0.4
